chore: update deps + option & future instrument kind
