+++ Cargo.toml
@@ -20,7 +20,7 @@ tracing = "0.1.29"
 # SerDe
 serde = { version = "1.0.143", features = ["derive"] }
 serde_json = "1.0.83"
-serde_qs = "0.11.0"
+serde_qs = "0.12.0"
 
 # Error
 thiserror = "1.0.32"
@@ -44,3 +44,4 @@ hex = "0.4.3"
 # Misc
 chrono = "0.4.21"
 bytes = "1.2.1"
+rust_decimal = "1.29.1"
