+++ src/model/mod.rs
@@ -1,8 +1,12 @@
+use crate::model::instrument::{kind::InstrumentKind, symbol::Symbol, Instrument};
 use serde::{Deserialize, Deserializer, Serialize};
 use std::{
     borrow::Cow,
     fmt::{Debug, Display, Formatter},
 };
+
+/// Todo:
+pub mod instrument;
 
 /// Represents a unique combination of an [`Exchange`] & an [`Instrument`].
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Deserialize, Serialize)]
@@ -123,130 +127,6 @@ impl Display for Exchange {
     }
 }
 
-/// Barter representation of an `Instrument`. Used to uniquely identify a `base_quote` pair, and it's
-/// associated instrument type.
-///
-/// eg/ Instrument { base: "btc", quote: "usdt", kind: Spot }
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Deserialize, Serialize)]
-pub struct Instrument {
-    pub base: Symbol,
-    pub quote: Symbol,
-    #[serde(rename = "instrument_type")]
-    pub kind: InstrumentKind,
-}
-
-impl Display for Instrument {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "({}_{}, {})", self.base, self.quote, self.kind)
-    }
-}
-
-impl<S> From<(S, S, InstrumentKind)> for Instrument
-where
-    S: Into<Symbol>,
-{
-    fn from((base, quote, kind): (S, S, InstrumentKind)) -> Self {
-        Self {
-            base: base.into(),
-            quote: quote.into(),
-            kind,
-        }
-    }
-}
-
-impl Instrument {
-    /// Constructs a new [`Instrument`] using the provided configuration.
-    pub fn new<S>(base: S, quote: S, kind: InstrumentKind) -> Self
-    where
-        S: Into<Symbol>,
-    {
-        Self {
-            base: base.into(),
-            quote: quote.into(),
-            kind,
-        }
-    }
-}
-
-/// Defines the type of [`Instrument`] which is being traded on a given `base_quote` market.
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum InstrumentKind {
-    Spot,
-    FuturePerpetual,
-}
-
-impl Default for InstrumentKind {
-    fn default() -> Self {
-        Self::Spot
-    }
-}
-
-impl Display for InstrumentKind {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                InstrumentKind::Spot => "spot",
-                InstrumentKind::FuturePerpetual => "future_perpetual",
-            }
-        )
-    }
-}
-
-/// Barter new type representing a currency symbol `String` identifier.
-///
-/// eg/ "btc", "eth", "usdt", etc
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize)]
-pub struct Symbol(String);
-
-impl Debug for Symbol {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Display for Symbol {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl AsRef<str> for Symbol {
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl<'de> Deserialize<'de> for Symbol {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        String::deserialize(deserializer).map(Symbol::new)
-    }
-}
-
-impl<S> From<S> for Symbol
-where
-    S: Into<String>,
-{
-    fn from(input: S) -> Self {
-        Symbol::new(input)
-    }
-}
-
-impl Symbol {
-    /// Construct a new [`Symbol`] new type using the provided `Into<Symbol>` value.
-    pub fn new<S>(input: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Self(input.into().to_lowercase())
-    }
-}
-
 /// New type representing a unique `String` identifier for a stream that has been subscribed to.
 /// This is used to identify data structures received over the socket.
 ///
@@ -313,10 +193,11 @@ impl Display for Side {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::model::instrument::{kind::InstrumentKind, Instrument};
     use serde::de::Error;
 
     #[test]
-    fn test_deserialise_market() {
+    fn test_de_market() {
         struct TestCase {
             input: &'static str,
             expected: Result<Market, serde_json::Error>,
@@ -353,63 +234,6 @@ mod tests {
 
         for (index, test) in cases.into_iter().enumerate() {
             let actual = serde_json::from_str::<Market>(test.input);
-
-            match (actual, test.expected) {
-                (Ok(actual), Ok(expected)) => {
-                    assert_eq!(actual, expected, "TC{} failed", index)
-                }
-                (Err(_), Err(_)) => {
-                    // Test passed
-                }
-                (actual, expected) => {
-                    // Test failed
-                    panic!("TC{index} failed because actual != expected. \nActual: {actual:?}\nExpected: {expected:?}\n");
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_deserialise_instrument() {
-        struct TestCase {
-            input: &'static str,
-            expected: Result<Instrument, serde_json::Error>,
-        }
-
-        let cases = vec![
-            TestCase {
-                // TC0: Valid Spot Instrument
-                input: r##"{"base": "btc", "quote": "usd", "instrument_type": "spot" }"##,
-                expected: Ok(Instrument::from(("btc", "usd", InstrumentKind::Spot))),
-            },
-            TestCase {
-                // TC1: Valid FuturePerpetual Instrument
-                input: r##"{"base": "btc", "quote": "usd", "instrument_type": "future_perpetual" }"##,
-                expected: Ok(Instrument::from((
-                    "btc",
-                    "usd",
-                    InstrumentKind::FuturePerpetual,
-                ))),
-            },
-            TestCase {
-                // TC2: Invalid Spot Instrument w/ numeric base
-                input: r##"{ "base": 100, "quote": "usd", "instrument_type": "future_perpetual" }"##,
-                expected: Err(serde_json::Error::custom("")),
-            },
-            TestCase {
-                // TC3: Invalid Instrument w/ gibberish InstrumentKind
-                input: r##"{"base": "btc", "quote": "usd", "instrument_type": "gibberish" }"##,
-                expected: Err(serde_json::Error::custom("")),
-            },
-            TestCase {
-                // TC4: Invalid Instrument w/ complete gibberish
-                input: r##"{ "gibberish": "shouldfail"}"##,
-                expected: Err(serde_json::Error::custom("")),
-            },
-        ];
-
-        for (index, test) in cases.into_iter().enumerate() {
-            let actual = serde_json::from_str::<Instrument>(test.input);
 
             match (actual, test.expected) {
                 (Ok(actual), Ok(expected)) => {
